third_party/WebKit/LayoutTests/editing/selection/extend/extend_selection_character_16.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script><script>
function testIt(dir, offset, direction, expectedText) {
  test(() => assert_selection(
  `<span dir="${dir}">${makeSample(offset)}</span>`,
  selection => selection.modify('extend', direction, 'character'),
  `<span dir="${dir}">${expectedText}</span>`),
  `${dir} ${offset} extend ${direction} character`
  );
}

function makeSample(offset){
  const kSample = 'car means \u05D0;\u05D1;\u05D2;.';
  return kSample.substr(0, offset) + '|' + kSample.substring(offset);
}

// ltr right character
testIt('ltr', 0, 'right', '^c|ar means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 1, 'right', 'c^a|r means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 2, 'right', 'ca^r| means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 3, 'right', 'car^ |means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 4, 'right', 'car ^m|eans \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 5, 'right', 'car m^e|ans \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 6, 'right', 'car me^a|ns \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 7, 'right', 'car mea^n|s \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 8, 'right', 'car mean^s| \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 9, 'right', 'car means^ |\u05D0;\u05D1;\u05D2;.');
testIt('ltr', 10, 'right', 'car means ^\u05D0|;\u05D1;\u05D2;.');
testIt('ltr', 11, 'right', 'car means \u05D0^;|\u05D1;\u05D2;.');
testIt('ltr', 12, 'right', 'car means \u05D0;^\u05D1|;\u05D2;.');
testIt('ltr', 13, 'right', 'car means \u05D0;\u05D1^;|\u05D2;.');
testIt('ltr', 14, 'right', 'car means \u05D0;\u05D1;^\u05D2|;.');
testIt('ltr', 15, 'right', 'car means \u05D0;\u05D1;\u05D2^;|.');
testIt('ltr', 16, 'right', 'car means \u05D0;\u05D1;\u05D2;^.|');

// ltr left character
testIt('ltr', 0, 'left', '|car means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 1, 'left', '|c^ar means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 2, 'left', 'c|a^r means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 3, 'left', 'ca|r^ means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 4, 'left', 'car| ^means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 5, 'left', 'car |m^eans \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 6, 'left', 'car m|e^ans \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 7, 'left', 'car me|a^ns \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 8, 'left', 'car mea|n^s \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 9, 'left', 'car mean|s^ \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 10, 'left', 'car means| ^\u05D0;\u05D1;\u05D2;.');
testIt('ltr', 11, 'left', 'car means |\u05D0^;\u05D1;\u05D2;.');
testIt('ltr', 12, 'left', 'car means \u05D0|;^\u05D1;\u05D2;.');
testIt('ltr', 13, 'left', 'car means \u05D0;|\u05D1^;\u05D2;.');
testIt('ltr', 14, 'left', 'car means \u05D0;\u05D1|;^\u05D2;.');
testIt('ltr', 15, 'left', 'car means \u05D0;\u05D1;|\u05D2^;.');
testIt('ltr', 16, 'left', 'car means \u05D0;\u05D1;\u05D2|;^.');

// ltr forward character
testIt('ltr', 0, 'forward', '^c|ar means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 1, 'forward', 'c^a|r means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 2, 'forward', 'ca^r| means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 3, 'forward', 'car^ |means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 4, 'forward', 'car ^m|eans \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 5, 'forward', 'car m^e|ans \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 6, 'forward', 'car me^a|ns \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 7, 'forward', 'car mea^n|s \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 8, 'forward', 'car mean^s| \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 9, 'forward', 'car means^ |\u05D0;\u05D1;\u05D2;.');
testIt('ltr', 10, 'forward', 'car means ^\u05D0|;\u05D1;\u05D2;.');
testIt('ltr', 11, 'forward', 'car means \u05D0^;|\u05D1;\u05D2;.');
testIt('ltr', 12, 'forward', 'car means \u05D0;^\u05D1|;\u05D2;.');
testIt('ltr', 13, 'forward', 'car means \u05D0;\u05D1^;|\u05D2;.');
testIt('ltr', 14, 'forward', 'car means \u05D0;\u05D1;^\u05D2|;.');
testIt('ltr', 15, 'forward', 'car means \u05D0;\u05D1;\u05D2^;|.');
testIt('ltr', 16, 'forward', 'car means \u05D0;\u05D1;\u05D2;^.|');

// ltr backward character
testIt('ltr', 0, 'backward', '|car means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 1, 'backward', '|c^ar means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 2, 'backward', 'c|a^r means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 3, 'backward', 'ca|r^ means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 4, 'backward', 'car| ^means \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 5, 'backward', 'car |m^eans \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 6, 'backward', 'car m|e^ans \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 7, 'backward', 'car me|a^ns \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 8, 'backward', 'car mea|n^s \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 9, 'backward', 'car mean|s^ \u05D0;\u05D1;\u05D2;.');
testIt('ltr', 10, 'backward', 'car means| ^\u05D0;\u05D1;\u05D2;.');
testIt('ltr', 11, 'backward', 'car means |\u05D0^;\u05D1;\u05D2;.');
testIt('ltr', 12, 'backward', 'car means \u05D0|;^\u05D1;\u05D2;.');
testIt('ltr', 13, 'backward', 'car means \u05D0;|\u05D1^;\u05D2;.');
testIt('ltr', 14, 'backward', 'car means \u05D0;\u05D1|;^\u05D2;.');
testIt('ltr', 15, 'backward', 'car means \u05D0;\u05D1;|\u05D2^;.');
testIt('ltr', 16, 'backward', 'car means \u05D0;\u05D1;\u05D2|;^.');

// rtl right character
testIt('rtl', 0, 'right', '^c|ar means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 1, 'right', 'c^a|r means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 2, 'right', 'ca^r| means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 3, 'right', 'car^ |means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 4, 'right', 'car ^m|eans \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 5, 'right', 'car m^e|ans \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 6, 'right', 'car me^a|ns \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 7, 'right', 'car mea^n|s \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 8, 'right', 'car mean^s| \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 9, 'right', 'car means^ |\u05D0;\u05D1;\u05D2;.');
testIt('rtl', 10, 'right', 'car means ^\u05D0|;\u05D1;\u05D2;.');
testIt('rtl', 11, 'right', 'car means \u05D0^;|\u05D1;\u05D2;.');
testIt('rtl', 12, 'right', 'car means \u05D0;^\u05D1|;\u05D2;.');
testIt('rtl', 13, 'right', 'car means \u05D0;\u05D1^;|\u05D2;.');
testIt('rtl', 14, 'right', 'car means \u05D0;\u05D1;^\u05D2|;.');
testIt('rtl', 15, 'right', 'car means \u05D0;\u05D1;\u05D2^;|.');
testIt('rtl', 16, 'right', 'car means \u05D0;\u05D1;\u05D2;^.|');

// rtl left character
testIt('rtl', 0, 'left', '|car means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 1, 'left', '|c^ar means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 2, 'left', 'c|a^r means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 3, 'left', 'ca|r^ means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 4, 'left', 'car| ^means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 5, 'left', 'car |m^eans \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 6, 'left', 'car m|e^ans \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 7, 'left', 'car me|a^ns \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 8, 'left', 'car mea|n^s \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 9, 'left', 'car mean|s^ \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 10, 'left', 'car means| ^\u05D0;\u05D1;\u05D2;.');
testIt('rtl', 11, 'left', 'car means |\u05D0^;\u05D1;\u05D2;.');
testIt('rtl', 12, 'left', 'car means \u05D0|;^\u05D1;\u05D2;.');
testIt('rtl', 13, 'left', 'car means \u05D0;|\u05D1^;\u05D2;.');
testIt('rtl', 14, 'left', 'car means \u05D0;\u05D1|;^\u05D2;.');
testIt('rtl', 15, 'left', 'car means \u05D0;\u05D1;|\u05D2^;.');
testIt('rtl', 16, 'left', 'car means \u05D0;\u05D1;\u05D2|;^.');

// rtl forward character
testIt('rtl', 0, 'forward', '^c|ar means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 1, 'forward', 'c^a|r means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 2, 'forward', 'ca^r| means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 3, 'forward', 'car^ |means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 4, 'forward', 'car ^m|eans \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 5, 'forward', 'car m^e|ans \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 6, 'forward', 'car me^a|ns \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 7, 'forward', 'car mea^n|s \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 8, 'forward', 'car mean^s| \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 9, 'forward', 'car means^ |\u05D0;\u05D1;\u05D2;.');
testIt('rtl', 10, 'forward', 'car means ^\u05D0|;\u05D1;\u05D2;.');
testIt('rtl', 11, 'forward', 'car means \u05D0^;|\u05D1;\u05D2;.');
testIt('rtl', 12, 'forward', 'car means \u05D0;^\u05D1|;\u05D2;.');
testIt('rtl', 13, 'forward', 'car means \u05D0;\u05D1^;|\u05D2;.');
testIt('rtl', 14, 'forward', 'car means \u05D0;\u05D1;^\u05D2|;.');
testIt('rtl', 15, 'forward', 'car means \u05D0;\u05D1;\u05D2^;|.');
testIt('rtl', 16, 'forward', 'car means \u05D0;\u05D1;\u05D2;^.|');

// rtl backward character
testIt('rtl', 0, 'backward', '|car means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 1, 'backward', '|c^ar means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 2, 'backward', 'c|a^r means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 3, 'backward', 'ca|r^ means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 4, 'backward', 'car| ^means \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 5, 'backward', 'car |m^eans \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 6, 'backward', 'car m|e^ans \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 7, 'backward', 'car me|a^ns \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 8, 'backward', 'car mea|n^s \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 9, 'backward', 'car mean|s^ \u05D0;\u05D1;\u05D2;.');
testIt('rtl', 10, 'backward', 'car means| ^\u05D0;\u05D1;\u05D2;.');
testIt('rtl', 11, 'backward', 'car means |\u05D0^;\u05D1;\u05D2;.');
testIt('rtl', 12, 'backward', 'car means \u05D0|;^\u05D1;\u05D2;.');
testIt('rtl', 13, 'backward', 'car means \u05D0;|\u05D1^;\u05D2;.');
testIt('rtl', 14, 'backward', 'car means \u05D0;\u05D1|;^\u05D2;.');
testIt('rtl', 15, 'backward', 'car means \u05D0;\u05D1;|\u05D2^;.');
testIt('rtl', 16, 'backward', 'car means \u05D0;\u05D1;\u05D2|;^.');
</script>
